requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pytest
httpx
passlib
python-jose